fixup! refactor: web-client webdav implementation
